fix neutrino_msg_data_t definition
neutrino_msg_data_t is used to pass pointers around, so it needs to
be the same length as a pointer. uint is not the same length as a
pointer on 64bit systems, so fix this up by defining it to size_t.

+++ tuxbox/neutrino/src/driver/rcinput.h
@@ -81,7 +81,7 @@
 
 
 typedef uint neutrino_msg_t;
-typedef uint neutrino_msg_data_t;
+typedef size_t neutrino_msg_data_t; // we are pushing pointer addresses around with this, so it needs to be long
 
 #define NEUTRINO_UDS_NAME "/tmp/neutrino.sock"
